adapt to multiple JS modules
